+++ R/eta_squared-methods.R
@@ -42,8 +42,9 @@
     insight::format_error("Cannot estimate approx effect size for `mixed` type model - no F-statistic found.")
   }
 
-  if (verbose && include_intercept) {
+  if (verbose && include_intercept && !"(Intercept)" %in% rownames(aov_tab)) {
     insight::format_warning("Cannot estimate (Intercept) effect size for `mixed` model.")
+    include_intercept <- FALSE
   }
 
   aov_tab$Parameter <- rownames(aov_tab)
@@ -51,7 +52,7 @@
   aov_tab$df_error <- aov_tab[["den Df"]]
   aov_tab <- aov_tab[, c("Parameter", "df", "df_error", "F")]
 
-  out <- .es_aov_table(aov_tab, verbose = verbose, ...)
+  out <- .es_aov_table(aov_tab, verbose = verbose, include_intercept = include_intercept, ...)
 
   attr(out, "anova_type") <- attr(model, "type")
   attr(out, "approximate") <- TRUE
@@ -405,22 +406,8 @@
 
 
 #' @export
-.anova_es.model_fit <- function(model,
-                                type = c("eta", "omega", "epsilon"),
-                                partial = TRUE,
-                                generalized = FALSE,
-                                ci = 0.95, alternative = "greater",
-                                verbose = TRUE,
-                                ...) {
-  .anova_es(
-    model$fit,
-    type = type,
-    partial = partial,
-    generalized = generalized,
-    ci = ci, alternative = alternative,
-    verbose = verbose,
-    ...
-  )
+.anova_es.model_fit <- function(model, ...) {
+  .anova_es(model$fit, ...)
 }